feat: Add lazy loading for images and update npm dependencies
This commit adds lazy loading for images in the `next.config.js` file, improving page load performance by loading images only when they are visible on the screen. It also updates the npm dependencies for `swr` to version 2.2.5. These changes enhance the user experience and ensure the application is using the latest stable dependencies.